refactor(weather): Move weather screens to feature-specific directory
- Relocated city_search_screen.dart, error_display_screen.dart, and weather_result_screen.dart
to features/weather/screens/ for improved modularity.
- Updated relative import paths for components and providers accordingly.
